refactor(BaseballNumberCache): change access level
- 야구게임숫자 캐시의 사이드이펙트를 막기 위해 변수 접근 권한을 변경했습니다.
- 이제 baseballNumber 패키지 외의 다른 위치에서는 해당 변수에 접근 할 수 없습니다.
