--- src/main/java/domain/BaseballNumberCache.java
@@ -1,8 +1,0 @@
-package domain;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class BaseballNumberCache {
-    public static final Map<Integer, BaseballNumber> lookup = new HashMap<>();
-}
