+++ src/main/java/domain/baseballNumber/BaseballNumberCache.java
@@ -1,0 +1,8 @@
+package domain.baseballNumber;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class BaseballNumberCache {
+    static final Map<Integer, BaseballNumber> lookup = new HashMap<>();
+}
