+++ src/test/java/domain/MockBaseballNumbersGeneratorTests.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.baseballNumber.BaseballNumber;
 import domain.exceptions.BaseballNumberDuplicatedException;
 import domain.exceptions.OutOfBoundBaseballNumberException;
 import org.junit.jupiter.api.DisplayName;
